Move Panels to own folder and linter cleanup
