+++ frontend/src/DeviceDetailsPage/TabPanels/DeviceDetailsActivityPanel.jsx
