+++ frontend/src/DeviceDetailsPage/TabPanels/DeviceDetailsNotesPanel.jsx
